refactor(agents): agents/ folder seam, pi's extension template first
ADR 0042 phase 3: create the host-internal agents/ home for
quirky-agent modules (as opposed to a thin catalog entry) and move
the first candidate into it.

agent-pi-extension.ts (pi's TypeScript hook-extension template,
renderPiTrackSessionExtension/PiExtensionParams) is already
self-contained and pi-only with a single consumer, so it moves to
agents/pi.ts unchanged. The old path becomes a one-line re-export so
agent-catalog.ts's existing import needs no change yet — that's the
seam. Phase 4 folds pi's AgentDefinition object (still in
agent-catalog.ts) into this same file and removes the shim.

pnpm test (1363 tests, same count) and lint pass unchanged.
